refactoring wrt. code review, bugfix rect. pattern
