+++ packages/shared/src/badukWithAbstractBoard/abstractBoard/StoneState.ts
