+++ packages/shared/src/badukWithAbstractBoard/abstractBoard/intersection.ts
